+++ app/components/dashboard/EntriesList.js
@@ -1,11 +1,12 @@
 import React, { PropTypes } from 'react';
+import moment from 'moment';
 
 const EntriesList = (props) => (
   <tbody>{ props.entries.map((entry, index) => {
     return (<tr key={index}>
       <td>{entry.title}</td>
-      <td>{entry.created}</td>
-      <td>{entry.updated}</td>
+      <td>{moment(entry.created).format('MM/DD/YYYY hh:mma')}</td>
+      <td>{entry.updated ? moment(entry.updated).format('MM/DD/YYYY hh:mma') : ''}</td>
       <td>{entry.category}</td>
     </tr>);
   })}